+++ tests/test_sync.py
@@ -8,7 +8,7 @@ from pip.download import path_to_url, url_to_path
 
 from prequ.exceptions import IncompatibleRequirements
 from prequ.sync import dependency_tree, diff, merge, sync
-from prequ.utils import get_ireq_version
+from prequ.utils import get_ireq_version, key_from_dist, normalize_req_name
 
 
 @pytest.mark.parametrize(
@@ -46,11 +46,12 @@ from prequ.utils import get_ireq_version
     ]
 )
 def test_dependency_tree(fake_dist, installed, root, expected):
-    installed = {distribution.key: distribution
+    installed = {key_from_dist(distribution): distribution
                  for distribution in
                  (fake_dist(name, deps) for name, deps in installed)}
 
-    actual = dependency_tree(installed, root)
+    root_key = normalize_req_name(root)
+    actual = dependency_tree(installed, root_key)
     assert actual == set(expected)